更改当前工作路径

+++ ch2/exp_2_1_mnist_mlp/main_exp_2_1.py
@@ -5,14 +5,14 @@ import numpy as np
 def evaluate(mlp):
     pred_results = np.zeros([mlp.test_data.shape[0]])
 
-    for idx in range(mlp.test_data.shape[0]/mlp.batch_size):
-        batch_images = mlp.test_data[idx * mlp.batch_size: (idx+1) * mlp.batch_size, :-1]
+    for idx in range(int(mlp.test_data.shape[0] / mlp.batch_size)):
+        batch_images = mlp.test_data[idx * mlp.batch_size: (idx + 1) * mlp.batch_size, :-1]
         prob = mlp.forward(batch_images)
         pred_labels = np.argmax(prob, axis=1)
-        pred_results[idx*mlp.batch_size:(idx+1)*mlp.batch_size] = pred_labels
+        pred_results[idx * mlp.batch_size:(idx + 1) * mlp.batch_size] = pred_labels
 
-    if mlp.test_data.shape[0] % mlp.batch_size >0: 
-        last_batch = mlp.test_data.shape[0]/mlp.batch_size*mlp.batch_size
+    if mlp.test_data.shape[0] % mlp.batch_size > 0:
+        last_batch = mlp.test_data.shape[0] / mlp.batch_size * mlp.batch_size
         batch_images = mlp.test_data[-last_batch:, :-1]
         prob = mlp.forward(batch_images)
         pred_labels = np.argmax(prob, axis=1)
